% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{A longitudinal example dataset.}
\format{
A data frame with 6700 rows and 9 variables:
\describe{
\item{id}{respondent id}
\item{wave}{wave number}
\item{score}{respondent score}
\item{weight}{sampling weight}
\item{strata}{strata variable}
\item{cluster}{cluster variable}
\item{ageg}{cathegorical age}
\item{ecacg}{educational level}
\item{qualifg}{economic activity}
}
}
\usage{
example_data
}
\description{
An example dataset containing the individuals scores for certain subject.
}
\keyword{datasets}
