% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.main.R
\name{mmcsd}
\alias{mmcsd}
\title{Fit fixed and random effects longitudinal model.}
\usage{
mmcsd(formula, waves, ids, weights, stratum, cluster, data, sigma = "identity")
}
\arguments{
\item{formula}{A formula}

\item{waves}{a dataframe column or an array}

\item{ids}{a dataframe column or an array}

\item{weights}{a dataframe column or an array}

\item{stratum}{a dataframe column or an array}

\item{cluster}{a dataframe column or an array}

\item{data}{A dataframe or tibble}

\item{sigma}{A character or a square matrix}
}
\value{
The fit model with class 'mmcsd'.
}
\description{
Estimate the fixed effects of the model, also known as B parameters of the regression,taking into account the sampling plan of the research, and also estimating the covariance matrix of the model considering the estimates of B
}
\examples{
\donttest{
fit <- mmcsd(
 score ~ wave + ageg + ecacg + qualifg,
 waves = wave, ids = id,
 weights = weight, stratum = strata, cluster = cluster,
 data = example_data, sigma = "exchangeable"
)
}
}
