% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.10.main.R
\name{sigmaThetaExpr_viewer}
\alias{sigmaThetaExpr_viewer}
\title{covariance structure viewer to preview sigmaThetaExpr to be used in 'cov_mmcsd'.}
\usage{
sigmaThetaExpr_viewer(sigmaThetaExpr, numWaves = NULL)
}
\arguments{
\item{sigmaThetaExpr}{A character with the covariance structure type or a list of expressions}

\item{numWaves}{An integer with the size of the square matrix to be printed.}
}
\value{
Return NULL and print in terminal the sigmaThetaExpr.
}
\description{
Knowing the difficulty of visualizing the covariance structure, especially when the user chooses to determine his own structure. This function was developed,that allows the user to view the provided structure even before it is evaluated, that is, through mathematics symbolic.
}
\examples{
sigmaThetaExpr_viewer("UCM", 5)
}
