% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.8.main.R
\name{summary.mmcsd}
\alias{summary.mmcsd}
\title{Summarise the results of 'mmcsd' fit.}
\usage{
\method{summary}{mmcsd}(object, ...)
}
\arguments{
\item{object}{A mmcsd fitted model}

\item{...}{Additional params passed to summary}
}
\value{
Return NULL and print in terminal the results.
}
\description{
Summarise the results of 'mmcsd' fit.
}
\examples{
\donttest{
fit <- mmcsd(
 score ~ wave + ageg + ecacg + qualifg,
 waves = wave, ids = id,
 weights = weight, stratum = strata, cluster = cluster,
 data = example_data, sigma = "exchangeable"
)
summary(fit)
}
}
