% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.10.main.R
\name{sigmaThetaExpr_viewer}
\alias{sigmaThetaExpr_viewer}
\title{covariance structure viewer to preview sigmaThetaExpr to be used in 'cov_mmcsd'.}
\usage{
sigmaThetaExpr_viewer(sigmaThetaExpr, numWaves = NULL)
}
\arguments{
\item{sigmaThetaExpr}{A character with the covariance structure type or a list of expressions}

\item{numWaves}{An integer with the size of the square matrix to be printed.}
}
\value{
Return NULL and print in terminal the sigmaThetaExpr.
}
\description{
covariance structure viewer to preview sigmaThetaExpr to be used in 'cov_mmcsd'.
}
