% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.11.main.R
\name{summary.mmcsd.theta}
\alias{summary.mmcsd.theta}
\title{Summarise the results of 'cov_mmcsd' fit.}
\usage{
\method{summary}{mmcsd.theta}(object, ...)
}
\arguments{
\item{object}{A mmcsd.theta fitted model}

\item{...}{Additional params passed to summary}
}
\value{
Return NULL and print in terminal the results.
}
\description{
Summarise the results of 'cov_mmcsd' fit.
}
